Update layout, move client side to layoutProvider
Also cleans page.tsx
