+++ src/app/layoutProvider.tsx
@@ -1,0 +1,109 @@
+"use client";
+
+import { CaretUpOutlined } from "@ant-design/icons";
+import { ConfigProvider, Layout, theme } from "antd";
+import Navbar from "Comp/Navbar";
+
+const { Header, Content, Footer } = Layout;
+
+/**
+ * returns the main html page as layout. Imports external JS and CSS libraries. Provides html head tags.
+ * @param children
+ * @returns html
+ */
+export default function RootLayoutProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  function goUp() {
+    document.body.scrollTop = 0; // For Safari
+    document.documentElement.scrollTop = 0; // For Chrome, Firefox, IE and Opera
+  }
+
+  return (
+    <ConfigProvider
+      theme={{
+        algorithm: theme.darkAlgorithm,
+      }}
+    >
+      <Layout style={{ minHeight: "100vh" }}>
+        <Header
+          style={{
+            padding: "0px",
+            position: "sticky",
+            top: 0,
+            zIndex: 1,
+            width: "100%",
+            display: "flex",
+            alignItems: "center",
+          }}
+        >
+          <Navbar />
+        </Header>
+        <Content
+          id="PageContent"
+          style={{ margin: "1rem auto 0 auto" }}
+          className="container"
+        >
+          {children}
+        </Content>
+        <Footer style={{ textAlign: "center" }}>
+          Teothe, Teothe3K, their information, their documents and the TTRPG
+          content is provided under the{" "}
+          <a
+            target="_blank"
+            href="https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode"
+          >
+            CC BY-NC-SA License{" "}
+          </a>
+          and protected by the{" "}
+          <a target="_blank" href="./SRD-OGL_V5.1.pdf">
+            Wizards OGL{" "}
+          </a>
+          and{" "}
+          <a target="_blank" href="./SRD5.1-CCBY4.0_License_livelinks.pdf">
+            Wizards CC BY
+          </a>
+          . Further attributions can be found on the{" "}
+          <a
+            target="_blank"
+            href="https://ozgurozbek.github.io/dnd/img-credits.xml"
+          >
+            Image Credits
+          </a>
+          ,{" "}
+          <a target="_blank" href="https://www.dndbeyond.com/">
+            D&D Beyond Website
+          </a>
+          ,{" "}
+          <a target="_blank" href="https://www.dmsguild.com/">
+            DM&apos;s Guild
+          </a>
+          ,{" "}
+          <a target="_blank" href="https://www.gmbinder.com/">
+            GM Binder
+          </a>
+          ,{" "}
+          <a target="_blank" href="https://www.dandwiki.com/wiki/">
+            D&D Wiki
+          </a>
+          , and{" "}
+          <a
+            target="_blank"
+            href="https://forgottenrealms.fandom.com/wiki/Main_Page"
+          >
+            Forgotten Realms Wiki
+          </a>
+          . 2016-2024.
+        </Footer>
+        <div
+          className="w-12 h-12 rounded-full bg-[#630436] hover:bg-[#30011a] fixed bottom-4 right-4 transition-all text-center text-4xl"
+          onClick={goUp}
+        >
+          <CaretUpOutlined />
+        </div>
+      </Layout>
+    </ConfigProvider>
+  );
+}
